refactor(worker-javascript): group extracted managers into src/core/
The 21 modules extracted from Core.js across PRs #1036-#1054
(`DiagnosticsManager`, `VisibilityTracker`, `StatePersistence`,
`AutoSubmitManager`, `NavigationHandler`, `ResolverAdapter`,
`RendererInstructionBuilder`, `ProcessQueue`, `ComponentLifecycle`,
`ChildMatcher`, `DeletionEngine`, `ActionTriggerScheduler`,
`StateVariableDefinitionFactory`, `StateVariableInitializer`,
`ComponentBuilder`, `CompositeExpander`, `EssentialValueWriter`,
`StateVariableEvaluator`, `StalenessPropagator`,
`CompositeReplacementUpdater`, `UpdateExecutor`) were sitting at the
root of `packages/doenetml-worker-javascript/src` alongside the Core
orchestrator itself, the proxy handlers, parser glue, and tests. With
21 sibling files at the same level, navigating to a particular manager
was getting noisy.

Move all 21 into `src/core/`, leaving `Core.ts` / `CoreWorker.ts` and
the unrelated infrastructure (`Numerics`, `ParameterStack`,
`Dependencies`, proxy handlers, `StateVariableNameResolver`) at the
root. Within-`core/` cross-imports stay relative (`./Foo`); imports
into `utils/` / `types/` / `Core` become `../utils/`, `../types/`,
`../Core`. The three external importers (`Core.ts`, `CoreWorker.ts`,
`Dependencies.js`) now go through `./core/Foo`.

`tsc --noEmit` reports the same set of errors as before the move
(63 pre-existing, all `: any`-style implicit-any complaints in the
StateVariableDefinitionFactory and CompositeExpander that the
in-progress typing pass tracks separately). Existing test groups
(`external_references`, `extend_references`, `baseComponentProperties`,
`calcStartEndIdx`) all pass.
