+++ packages/doenetml-worker-javascript/src/Core.ts
@@ -16,20 +16,20 @@ import { reportTimerError, TimerLabels } from "./utils/timerErrors";
 import { getSourceLocationForComponent } from "./utils/sourceLocation";
 import type { ComponentInstance } from "./types/componentInstance";
 import { DependencyHandler } from "./Dependencies";
-import { ActionTriggerScheduler } from "./ActionTriggerScheduler";
-import { AutoSubmitManager } from "./AutoSubmitManager";
-import { addComponents } from "./ComponentBuilder";
-import { CompositeReplacementUpdater } from "./CompositeReplacementUpdater";
-import { DiagnosticsManager } from "./DiagnosticsManager";
-import { EssentialValueWriter } from "./EssentialValueWriter";
-import { navigateToTarget } from "./NavigationHandler";
-import { ProcessQueue } from "./ProcessQueue";
-import { RendererInstructionBuilder } from "./RendererInstructionBuilder";
-import { StalenessPropagator } from "./StalenessPropagator";
-import { StatePersistence } from "./StatePersistence";
-import { StateVariableEvaluator } from "./StateVariableEvaluator";
-import { UpdateExecutor } from "./UpdateExecutor";
-import { VisibilityTracker } from "./VisibilityTracker";
+import { ActionTriggerScheduler } from "./core/ActionTriggerScheduler";
+import { AutoSubmitManager } from "./core/AutoSubmitManager";
+import { addComponents } from "./core/ComponentBuilder";
+import { CompositeReplacementUpdater } from "./core/CompositeReplacementUpdater";
+import { DiagnosticsManager } from "./core/DiagnosticsManager";
+import { EssentialValueWriter } from "./core/EssentialValueWriter";
+import { navigateToTarget } from "./core/NavigationHandler";
+import { ProcessQueue } from "./core/ProcessQueue";
+import { RendererInstructionBuilder } from "./core/RendererInstructionBuilder";
+import { StalenessPropagator } from "./core/StalenessPropagator";
+import { StatePersistence } from "./core/StatePersistence";
+import { StateVariableEvaluator } from "./core/StateVariableEvaluator";
+import { UpdateExecutor } from "./core/UpdateExecutor";
+import { VisibilityTracker } from "./core/VisibilityTracker";
 import * as nameResolver from "./StateVariableNameResolver";
 import { numberAnswers } from "./utils/answer";
