+++ packages/doenetml-worker-javascript/src/core/DiagnosticsManager.ts
