+++ packages/doenetml-worker-javascript/src/core/UpdateExecutor.ts
@@ -1,9 +1,9 @@
-import type Core from "./Core";
+import type Core from "../Core";
 import type { ComponentIdx } from "@doenet/utils";
 import { addComponents } from "./ComponentBuilder";
 import { deleteComponents } from "./DeletionEngine";
-import { createNewComponentIndices } from "./utils/componentIndices";
-import { reportTimerError, TimerLabels } from "./utils/timerErrors";
+import { createNewComponentIndices } from "../utils/componentIndices";
+import { reportTimerError, TimerLabels } from "../utils/timerErrors";
 
 /**
  * Source-side metadata about *how* an update originated. Indexed by
